requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
retell-sdk==4.0.0
rich==13.7.1
rpds-py==0.19.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
